factor out binary dictionary writer/reader for kuromoji
