+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/dict/JaMorphAttributes.java
@@ -16,39 +16,16 @@
  */
 package org.apache.lucene.analysis.ja.dict;
 
-/** Dictionary interface for retrieving morphological data by id. */
-public interface Dictionary {
+import org.apache.lucene.analysis.morph.MorphAttributes;
 
-  public static final String INTERNAL_SEPARATOR = "\u0000";
-
-  /**
-   * Get left id of specified word
-   *
-   * @return left id
-   */
-  public int getLeftId(int wordId);
-
-  /**
-   * Get right id of specified word
-   *
-   * @return right id
-   */
-  public int getRightId(int wordId);
-
-  /**
-   * Get word cost of specified word
-   *
-   * @return word's cost
-   */
-  public int getWordCost(int wordId);
-
+public interface JaMorphAttributes extends MorphAttributes {
   /**
    * Get Part-Of-Speech of tokens
    *
    * @param wordId word ID of token
    * @return Part-Of-Speech of the token
    */
-  public String getPartOfSpeech(int wordId);
+  String getPartOfSpeech(int wordId);
 
   /**
    * Get reading of tokens
@@ -56,7 +33,7 @@ public interface Dictionary {
    * @param wordId word ID of token
    * @return Reading of the token
    */
-  public String getReading(int wordId, char[] surface, int off, int len);
+  String getReading(int wordId, char[] surface, int off, int len);
 
   /**
    * Get base form of word
@@ -64,7 +41,7 @@ public interface Dictionary {
    * @param wordId word ID of token
    * @return Base form (only different for inflected words, otherwise null)
    */
-  public String getBaseForm(int wordId, char[] surface, int off, int len);
+  String getBaseForm(int wordId, char[] surface, int off, int len);
 
   /**
    * Get pronunciation of tokens
@@ -72,7 +49,7 @@ public interface Dictionary {
    * @param wordId word ID of token
    * @return Pronunciation of the token
    */
-  public String getPronunciation(int wordId, char[] surface, int off, int len);
+  String getPronunciation(int wordId, char[] surface, int off, int len);
 
   /**
    * Get inflection type of tokens
@@ -80,7 +57,7 @@ public interface Dictionary {
    * @param wordId word ID of token
    * @return inflection type, or null
    */
-  public String getInflectionType(int wordId);
+  String getInflectionType(int wordId);
 
   /**
    * Get inflection form of tokens
@@ -88,7 +65,7 @@ public interface Dictionary {
    * @param wordId word ID of token
    * @return inflection form, or null
    */
-  public String getInflectionForm(int wordId);
+  String getInflectionForm(int wordId);
   // TODO: maybe we should have a optimal method, a non-typesafe
   // 'getAdditionalData' if other dictionaries like unidic have additional data
 }
