+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/dict/TokenInfoDictionary.java
@@ -21,6 +21,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+
+import org.apache.lucene.analysis.morph.BinaryDictionary;
 import org.apache.lucene.store.DataInput;
 import org.apache.lucene.store.InputStreamDataInput;
 import org.apache.lucene.util.IOSupplier;
@@ -32,11 +34,12 @@ import org.apache.lucene.util.fst.PositiveIntOutputs;
  * Binary dictionary implementation for a known-word dictionary model: Words are encoded into an FST
  * mapping to a list of wordIDs.
  */
-public final class TokenInfoDictionary extends BinaryDictionary {
+public final class TokenInfoDictionary extends BinaryDictionary<TokenInfoMorphAttributes> {
 
   public static final String FST_FILENAME_SUFFIX = "$fst.dat";
 
   private final TokenInfoFST fst;
+  private final TokenInfoMorphAttributes morphAtts;
 
   /**
    * Create a {@link TokenInfoDictionary} from an external resource path.
@@ -70,7 +73,9 @@ public final class TokenInfoDictionary extends BinaryDictionary {
       IOSupplier<InputStream> dictResource,
       IOSupplier<InputStream> fstResource)
       throws IOException {
-    super(targetMapResource, posResource, dictResource);
+    super(targetMapResource, dictResource, Constants.TARGETMAP_HEADER, Constants.DICT_HEADER, Constants.VERSION);
+    this.morphAtts = new TokenInfoMorphAttributes(buffer, posResource);
+
     FST<Long> fst;
     try (InputStream is = new BufferedInputStream(fstResource.get())) {
       DataInput in = new InputStreamDataInput(is);
@@ -84,6 +89,11 @@ public final class TokenInfoDictionary extends BinaryDictionary {
     final String resourcePath = TokenInfoDictionary.class.getSimpleName() + suffix;
     return IOUtils.requireResourceNonNull(
         TokenInfoDictionary.class.getResourceAsStream(resourcePath), resourcePath);
+  }
+
+  @Override
+  public TokenInfoMorphAttributes getMorphAttributes() {
+    return morphAtts;
   }
 
   public TokenInfoFST getFST() {
