+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/dict/UnknownDictionary.java
@@ -20,12 +20,16 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+
+import org.apache.lucene.analysis.morph.BinaryDictionary;
+import org.apache.lucene.util.IOSupplier;
 import org.apache.lucene.util.IOUtils;
 
 /** Dictionary for unknown-word handling. */
-public final class UnknownDictionary extends BinaryDictionary {
+public final class UnknownDictionary extends BinaryDictionary<UnknownMorphAttributes> {
 
   private final CharacterDefinition characterDefinition = CharacterDefinition.getInstance();
+  private final UnknownMorphAttributes morphAtts;
 
   /**
    * Create a {@link UnknownDictionary} from an external resource path.
@@ -36,23 +40,36 @@ public final class UnknownDictionary extends BinaryDictionary {
    * @throws IOException if resource was not found or broken
    */
   public UnknownDictionary(Path targetMapFile, Path posDictFile, Path dictFile) throws IOException {
-    super(
+    this(
         () -> Files.newInputStream(targetMapFile),
         () -> Files.newInputStream(posDictFile),
         () -> Files.newInputStream(dictFile));
   }
 
   private UnknownDictionary() throws IOException {
-    super(
+    this(
         () -> getClassResource(TARGETMAP_FILENAME_SUFFIX),
         () -> getClassResource(POSDICT_FILENAME_SUFFIX),
         () -> getClassResource(DICT_FILENAME_SUFFIX));
+  }
+
+  private UnknownDictionary(
+    IOSupplier<InputStream> targetMapResource,
+    IOSupplier<InputStream> posResource,
+    IOSupplier<InputStream> dictResource) throws IOException {
+    super(targetMapResource, dictResource, Constants.TARGETMAP_HEADER, Constants.DICT_HEADER, Constants.VERSION);
+    this.morphAtts = new UnknownMorphAttributes(buffer, posResource);
   }
 
   private static InputStream getClassResource(String suffix) throws IOException {
     final String resourcePath = UnknownDictionary.class.getSimpleName() + suffix;
     return IOUtils.requireResourceNonNull(
         UnknownDictionary.class.getResourceAsStream(resourcePath), resourcePath);
+  }
+
+  @Override
+  public UnknownMorphAttributes getMorphAttributes() {
+    return morphAtts;
   }
 
   public int lookup(char[] text, int offset, int len) {
@@ -77,21 +94,6 @@ public final class UnknownDictionary extends BinaryDictionary {
 
   public CharacterDefinition getCharacterDefinition() {
     return characterDefinition;
-  }
-
-  @Override
-  public String getReading(int wordId, char[] surface, int off, int len) {
-    return null;
-  }
-
-  @Override
-  public String getInflectionType(int wordId) {
-    return null;
-  }
-
-  @Override
-  public String getInflectionForm(int wordId) {
-    return null;
   }
 
   public static UnknownDictionary getInstance() {
