+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/dict/UserDictionary.java
@@ -26,13 +26,16 @@ import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 import org.apache.lucene.analysis.ja.util.CSVUtil;
+import org.apache.lucene.analysis.morph.Dictionary;
 import org.apache.lucene.util.IntsRefBuilder;
 import org.apache.lucene.util.fst.FST;
 import org.apache.lucene.util.fst.FSTCompiler;
 import org.apache.lucene.util.fst.PositiveIntOutputs;
 
 /** Class for building a User Dictionary. This class allows for custom segmentation of phrases. */
-public final class UserDictionary implements Dictionary {
+public final class UserDictionary implements Dictionary<UserMorphAttributes> {
+
+  public static final String INTERNAL_SEPARATOR = "\u0000";
 
   // phrase text -> phrase ID
   private final TokenInfoFST fst;
@@ -41,15 +44,9 @@ public final class UserDictionary implements Dictionary {
   private final int[][] segmentations;
 
   // holds readings and POS, indexed by wordid
-  private final String[] data;
+  private final UserMorphAttributes morphAtts;
 
-  private static final int CUSTOM_DICTIONARY_WORD_ID_OFFSET = 100000000;
-
-  public static final int WORD_COST = -100000;
-
-  public static final int LEFT_ID = 5;
-
-  public static final int RIGHT_ID = 5;
+  static final int CUSTOM_DICTIONARY_WORD_ID_OFFSET = 100000000;
 
   public static UserDictionary open(Reader reader) throws IOException {
 
@@ -150,8 +147,13 @@ public final class UserDictionary implements Dictionary {
       ord++;
     }
     this.fst = new TokenInfoFST(fstCompiler.compile(), false);
-    this.data = data.toArray(new String[data.size()]);
+    //this.data = data.toArray(new String[data.size()]);
+    this.morphAtts = new UserMorphAttributes(data.toArray(new String[0]));
     this.segmentations = segmentations.toArray(new int[segmentations.size()][]);
+  }
+
+  public UserMorphAttributes getMorphAttributes() {
+    return morphAtts;
   }
 
   /**
@@ -221,79 +223,5 @@ public final class UserDictionary implements Dictionary {
 
   public int[] lookupSegmentation(int phraseID) {
     return segmentations[phraseID];
-  }
-
-  @Override
-  public int getLeftId(int wordId) {
-    return LEFT_ID;
-  }
-
-  @Override
-  public int getRightId(int wordId) {
-    return RIGHT_ID;
-  }
-
-  @Override
-  public int getWordCost(int wordId) {
-    return WORD_COST;
-  }
-
-  @Override
-  public String getReading(int wordId, char[] surface, int off, int len) {
-    return getFeature(wordId, 0);
-  }
-
-  @Override
-  public String getPartOfSpeech(int wordId) {
-    return getFeature(wordId, 1);
-  }
-
-  @Override
-  public String getBaseForm(int wordId, char[] surface, int off, int len) {
-    return null; // TODO: add support?
-  }
-
-  @Override
-  public String getPronunciation(int wordId, char[] surface, int off, int len) {
-    return null; // TODO: add support?
-  }
-
-  @Override
-  public String getInflectionType(int wordId) {
-    return null; // TODO: add support?
-  }
-
-  @Override
-  public String getInflectionForm(int wordId) {
-    return null; // TODO: add support?
-  }
-
-  private String[] getAllFeaturesArray(int wordId) {
-    String allFeatures = data[wordId - CUSTOM_DICTIONARY_WORD_ID_OFFSET];
-    if (allFeatures == null) {
-      return null;
-    }
-
-    return allFeatures.split(INTERNAL_SEPARATOR);
-  }
-
-  private String getFeature(int wordId, int... fields) {
-    String[] allFeatures = getAllFeaturesArray(wordId);
-    if (allFeatures == null) {
-      return null;
-    }
-    StringBuilder sb = new StringBuilder();
-    if (fields.length == 0) { // All features
-      for (String feature : allFeatures) {
-        sb.append(CSVUtil.quoteEscape(feature)).append(",");
-      }
-    } else if (fields.length == 1) { // One feature doesn't need to escape value
-      sb.append(allFeatures[fields[0]]).append(",");
-    } else {
-      for (int field : fields) {
-        sb.append(CSVUtil.quoteEscape(allFeatures[field])).append(",");
-      }
-    }
-    return sb.deleteCharAt(sb.length() - 1).toString();
   }
 }
