+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/util/UnknownDictionaryWriter.java
@@ -19,19 +19,21 @@ package org.apache.lucene.analysis.ja.util;
 import java.io.IOException;
 import java.nio.file.Path;
 import org.apache.lucene.analysis.ja.dict.CharacterDefinition;
+import org.apache.lucene.analysis.ja.dict.Constants;
 import org.apache.lucene.analysis.ja.dict.UnknownDictionary;
 
-class UnknownDictionaryWriter extends BinaryDictionaryWriter {
+class UnknownDictionaryWriter extends org.apache.lucene.analysis.morph.BinaryDictionaryWriter<UnknownDictionary> {
   private final CharacterDefinitionWriter characterDefinition = new CharacterDefinitionWriter();
 
   public UnknownDictionaryWriter(int size) {
-    super(UnknownDictionary.class, size);
+    super(UnknownDictionary.class, new TokenInfoDictionaryEntryWriter(size));
   }
 
   @Override
   public int put(String[] entry) {
     // Get wordId of current entry
-    int wordId = buffer.position();
+    //int wordId = buffer.position();
+    int wordId = entryWriter.currentPosition();
 
     // Put entry
     int result = super.put(entry);
@@ -56,9 +58,9 @@ class UnknownDictionaryWriter extends BinaryDictionaryWriter {
     characterDefinition.putInvokeDefinition(characterClassName, invoke, group, length);
   }
 
-  @Override
+  //@Override
   public void write(Path baseDir) throws IOException {
-    super.write(baseDir);
+    super.write(baseDir, Constants.TARGETMAP_HEADER, Constants.POSDICT_HEADER, Constants.DICT_HEADER, Constants.VERSION);
     characterDefinition.write(baseDir);
   }
 }
