+++ lucene/analysis/kuromoji/src/test/org/apache/lucene/analysis/ja/dict/TestTokenInfoDictionary.java
@@ -16,10 +16,10 @@
  */
 package org.apache.lucene.analysis.ja.dict;
 
-import static org.apache.lucene.analysis.ja.dict.BinaryDictionary.DICT_FILENAME_SUFFIX;
-import static org.apache.lucene.analysis.ja.dict.BinaryDictionary.POSDICT_FILENAME_SUFFIX;
-import static org.apache.lucene.analysis.ja.dict.BinaryDictionary.TARGETMAP_FILENAME_SUFFIX;
 import static org.apache.lucene.analysis.ja.dict.TokenInfoDictionary.FST_FILENAME_SUFFIX;
+import static org.apache.lucene.analysis.morph.BinaryDictionary.DICT_FILENAME_SUFFIX;
+import static org.apache.lucene.analysis.morph.BinaryDictionary.POSDICT_FILENAME_SUFFIX;
+import static org.apache.lucene.analysis.morph.BinaryDictionary.TARGETMAP_FILENAME_SUFFIX;
 
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
@@ -137,17 +137,17 @@ public class TestTokenInfoDictionary extends LuceneTestCase {
         assertTrue(wordId > lastWordId);
         lastWordId = wordId;
 
-        String baseForm = tid.getBaseForm(wordId, chars, 0, chars.length);
+        String baseForm = tid.getMorphAttributes().getBaseForm(wordId, chars, 0, chars.length);
         assertTrue(baseForm == null || UnicodeUtil.validUTF16String(baseForm));
 
-        String inflectionForm = tid.getInflectionForm(wordId);
+        String inflectionForm = tid.getMorphAttributes().getInflectionForm(wordId);
         assertTrue(inflectionForm == null || UnicodeUtil.validUTF16String(inflectionForm));
         if (inflectionForm != null) {
           // check that it's actually an ipadic inflection form
           assertNotNull(ToStringUtil.getInflectedFormTranslation(inflectionForm));
         }
 
-        String inflectionType = tid.getInflectionType(wordId);
+        String inflectionType = tid.getMorphAttributes().getInflectionType(wordId);
         assertTrue(inflectionType == null || UnicodeUtil.validUTF16String(inflectionType));
         if (inflectionType != null) {
           // check that it's actually an ipadic inflection type
@@ -161,17 +161,17 @@ public class TestTokenInfoDictionary extends LuceneTestCase {
 
         tid.getWordCost(wordId);
 
-        String pos = tid.getPartOfSpeech(wordId);
+        String pos = tid.getMorphAttributes().getPartOfSpeech(wordId);
         assertNotNull(pos);
         assertTrue(UnicodeUtil.validUTF16String(pos));
         // check that it's actually an ipadic pos tag
         assertNotNull(ToStringUtil.getPOSTranslation(pos));
 
-        String pronunciation = tid.getPronunciation(wordId, chars, 0, chars.length);
+        String pronunciation = tid.getMorphAttributes().getPronunciation(wordId, chars, 0, chars.length);
         assertNotNull(pronunciation);
         assertTrue(UnicodeUtil.validUTF16String(pronunciation));
 
-        String reading = tid.getReading(wordId, chars, 0, chars.length);
+        String reading = tid.getMorphAttributes().getReading(wordId, chars, 0, chars.length);
         assertNotNull(reading);
         assertTrue(UnicodeUtil.validUTF16String(reading));
       }
